Add repeated/map fields to test proto Fix bug causing repeated fields to not be added to declaration Split transpiler test to separate test file Fix typo in Declare docstring
